graph: use node IDs for retrieval and nodes for creation in graphs
Note, this commit does not yet compile, but is in an intermediate
state. The intention is to continue working on it within the next
few weeks, and to open up for a discussion of the API; specifically
regarding if we wish to introduce a dedicated graph.NodeID type, or
simply use int64 to distinguish node IDs. Both approaches have
benefits and drawbacks. Lets discuss them in the gonum/graph#87
issue.